add readme, add comparator to Sorter, delete useless exception

+++ src/main/java/com/custis/sorter/generator/Generator.java
@@ -4,5 +4,5 @@ import java.io.OutputStream;
 
 public interface Generator {
 
-    void fillFile(int countOfStrokes, int maxLength, OutputStream out);
+    void fillFile(long countOfStrokes, int maxLength, OutputStream out);
 }
